requirements.txt
strands-agents
strands-agents-tools
strands-agents[ollama]
boto3